refactor(twenty-server): apply review nits on encryption envelope PR
- Rename secret-encryption.constants.ts to .constant.ts (Charles).
- Replace `=== null` / `!== null` with isDefined typeguards (Charles).
- Rephrase the AES-CTR no-integrity comment to drop informal "garbage"
  wording (Charles).
