+++ packages/twenty-server/src/engine/core-modules/secret-encryption/utils/__tests__/parse-secret-encryption-envelope-or-throw.util.spec.ts
@@ -1,7 +1,7 @@
 import {
   SECRET_ENCRYPTION_ENVELOPE_V1_PREFIX,
   SECRET_ENCRYPTION_ENVELOPE_V2_PREFIX,
-} from 'src/engine/core-modules/secret-encryption/constants/secret-encryption.constants';
+} from 'src/engine/core-modules/secret-encryption/constants/secret-encryption.constant';
 import {
   SecretEncryptionException,
   SecretEncryptionExceptionCode,
