+++ packages/twenty-server/src/engine/core-modules/secret-encryption/utils/format-secret-encryption-envelope-v2.util.ts
@@ -1,4 +1,4 @@
-import { SECRET_ENCRYPTION_ENVELOPE_V2_PREFIX } from 'src/engine/core-modules/secret-encryption/constants/secret-encryption.constants';
+import { SECRET_ENCRYPTION_ENVELOPE_V2_PREFIX } from 'src/engine/core-modules/secret-encryption/constants/secret-encryption.constant';
 
 export const formatSecretEncryptionEnvelopeV2 = ({
   keyId,
