+++ packages/twenty-server/src/engine/core-modules/secret-encryption/utils/pick-encryption-key-by-key-id-or-throw.util.ts
@@ -1,3 +1,5 @@
+import { isDefined } from 'twenty-shared/utils';
+
 import {
   SecretEncryptionException,
   SecretEncryptionExceptionCode,
@@ -17,7 +19,7 @@ export const pickEncryptionKeyByKeyIdOrThrow = ({
   }
 
   if (
-    keys.fallback !== null &&
+    isDefined(keys.fallback) &&
     computeEncryptionKeyId({ rawKey: keys.fallback }) === keyId
   ) {
     return keys.fallback;
